requirements.txt
# Python 3.12+ Requirements for WhatsApp Multi-Agent System
# Consolidates all dependencies for production deployment

# ============================================
# Core Dependencies
# ============================================

# Claude Agent SDK (official Python SDK for building agents)
claude-agent-sdk>=0.1.0

# Anthropic Claude API (for low-level API access)
anthropic>=0.18.0

# HTTP Requests (for WhatsApp API calls)
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# ============================================
# Web Service Dependencies (Render Deployment)
# ============================================

# FastAPI for REST API
fastapi>=0.104.0

# ASGI server
uvicorn[standard]>=0.24.0

# ============================================
# MCP Server Dependencies
# ============================================

# Model Context Protocol SDK
mcp>=1.18.0

# ============================================
# Session Storage (Redis)
# ============================================

# Redis client for session persistence
redis>=5.0.0

# ============================================
# Database (PostgreSQL/Neon)
# ============================================

# SQLAlchemy ORM with async support
sqlalchemy[asyncio]>=2.0.0

# PostgreSQL async driver (works with Neon)
asyncpg>=0.29.0

# Alembic for database migrations
alembic>=1.13.0

# PostgreSQL MCP Server (for agent database access via MCP)
pgsql-mcp-server>=1.3.2

# ============================================
# Multi-Agent A2A Dependencies (Phase 1)
# ============================================

# Async HTTP client for A2A protocol
aiohttp>=3.9.0

# Type hints extensions
typing-extensions>=4.14.1

# ============================================
# Telemetry & Observability (Phase 2)
# ============================================

# Logfire - Pydantic's observability platform
logfire>=0.40.0
